requirements.txt
python-telegram-bot>=22.5
openai>=2.15.0
python-dotenv>=1.2.0
pypdf>=6.6.0
langchain>=1.2.6
langchain-openai>=1.1.7
langchain-community>=0.4.1
langchain-text-splitters>=1.1.0
chromadb>=1.4.1
tiktoken>=0.12.0
